requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.2
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
etelemetry==0.3.1
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fitz==0.0.1.dev2
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lightning-utilities==0.11.5
looseversion==1.3.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
prov==2.0.1
psutil==6.0.0
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2024.1
pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
rdflib==6.3.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scipy==1.14.0
setuptools==70.3.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
supervision==0.3.0
sympy==1.13.0
tbb==2021.13.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
transformers==4.42.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
